requirements.txt
Flask==2.2.2
torch==2.1.0
transformers==4.29.0
Pillow==9.5.0
facenet-pytorch==2.5.2
rembg==2.0.28
onnxruntime==1.15.0
opencv-python-headless==4.7.0.72
matplotlib==3.7.1
torchvision==0.15.2
